make pedestal plot also work for LCLS2
